packages/python/pyproject.toml
[tool.poetry]
name = "epistemic-edge"
version = "0.0.1"
description = "Air-gapped neuro-symbolic AIoT framework: 1-bit LLM cognition over mathematically verified, temporally valid edge state."
authors = ["Muntaser Syed <jemsbhai@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/jemsbhai/epistemic-edge"
repository = "https://github.com/jemsbhai/epistemic-edge"
keywords = [
    "aiot", "edge-ai", "1-bit-llm", "neuro-symbolic",
    "subjective-logic", "prov-o", "cbor-ld", "iot",
    "temporal-reasoning", "constrained-decoding",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet",
    "Topic :: System :: Hardware",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
packages = [{include = "epistemic_edge", from = "src"}]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/jemsbhai/epistemic-edge/issues"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
jsonld-ex = ">=0.7.0"
cbor-ld-ex = ">=0.2.0"
chronofy = ">=0.1.8"
pydantic = ">=2.0"

[tool.poetry.group.transport.dependencies]
aiomqtt = ">=2.0.0"
aiocoap = ">=0.4.7"

[tool.poetry.group.llm.dependencies]
llama-cpp-python = ">=0.2.56"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
pytest-cov = ">=5.0"
hypothesis = ">=6.100"
mypy = ">=1.10"
ruff = ">=0.4"

[tool.poetry.extras]
transport = ["aiomqtt", "aiocoap"]
llm = ["llama-cpp-python"]
all = ["aiomqtt", "aiocoap", "llama-cpp-python"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
